Created folder and .h for heredoc
